# docs/swagger.yaml
definitions:
...
  api.createUserRequest:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
  api.loginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.loginUserResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/api.userResponse'
    type: object
  api.userResponse:
    properties:
      name:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Simple golang blog for backend practice
  title: Golang-blog project
  version: "1.0"
paths:
...
  /api/users:
    post:
      consumes:
      - application/json
      description: Create user by username, password and name
      parameters:
      - description: register request
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      summary: Create a user
      tags:
      - user
  /api/users/login:
    post:
      consumes:
      - application/json
      description: login user by username and password
      parameters:
      - description: login request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginUserResponse'
      summary: login a user
      tags:
      - user
swagger: "2.0"